requirements.txt
#for api
fastapi
uvicorn

#for code cleaning
pre-commit

#for template
jinja2

#for static files
aiofiles

#for file processing
python-multipart

python-dotenv
pydantic
deta
